Pull out ErrorAlert further from Banner to help highlight this section
