+++ services/app-web/src/components/ErrorAlert/ErrorAlert.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import styles from '../Banner.module.scss'
+import styles from './ErrorAlert.module.scss'
 import { Alert, Link } from '@trussworks/react-uswds'
-import { MAIL_TO_SUPPORT } from '../../../constants/errors'
+import { MAIL_TO_SUPPORT } from '../../constants/errors'
 export type ErrorAlertProps = {
     message?: string | React.ReactNode
     heading?: string
