+++ services/app-web/src/pages/App/AppRoutes.tsx
@@ -15,7 +15,7 @@ import { CognitoLogin } from '../Auth/CognitoLogin'
 import { CMSDashboard } from '../CMSDashboard/CMSDashboard'
 import { StateDashboard } from '../StateDashboard/StateDashboard'
 import { AuthenticatedRouteWrapper } from '../Wrapper/AuthenticatedRouteWrapper'
-import { Error404 } from '../Errors/Error404'
+import { Error404 } from '../Errors/Error404Page'
 import { Help } from '../Help/Help'
 import { Landing } from '../Landing/Landing'
 import { NewStateSubmissionForm, StateSubmissionForm } from '../StateSubmission'
