+++ services/app-web/src/pages/Errors/ErrorBoundaryRoot.tsx
@@ -3,8 +3,12 @@ import { recordJSException, recordSpan } from '../../otelHelpers/tracingHelper'
 import { useEffect } from 'react'
 import { EmptyHeader } from '../../components/Header/EmptyHeader'
 import styles from '../App/AppBody.module.scss'
-// It is possible to have different error boundary fallback components at different levels of the hierarchy.
-// This is the root error boundary. If it is visible, this means that the React frontend has crashed entirely and the entire tree is gone.
+/* The error boundary is a fallback UI to catch errors
+    See React docs: https://reactjs.org/docs/error-boundaries.html
+    It is possible to have different error boundary fallback components at different levels of the hierarchy.
+    This component is the root error boundary. I visible, this means that the React frontend has crashed fully and the entire tree is gone. 
+    Accordingly, this page brings back in an empty header (with no login or stateful data) and the generic error page text.
+*/
 
 function ErrorBoundaryRoot({
     error,
